Add custom default command name and extension
This patch allows the customisation of the default command module name and the extension to use for lookups. Depending on the users build they may be using a different extensions for their CommonJS modules. Being able to specify the default command module name allows the user to re-use a nested sub-command as the default command.